Squashed commit of the following:
commit c79c3630c6427ac994525245b3229a945f5e0171
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Wed Jan 25 18:40:08 2023 +0100

    Revert "Revert "Revert "Revert "Revert "Revert "readme update""""""

    This reverts commit 7075080e50b95ec57dcf2fec2325d3ffe8e6178a.

commit 7075080e50b95ec57dcf2fec2325d3ffe8e6178a
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Wed Jan 25 18:40:04 2023 +0100

    Revert "Revert "Revert "Revert "Revert "readme update"""""

    This reverts commit 5643d73faebf9458afff4153ea27ecedfd43dc1e.

commit 5643d73faebf9458afff4153ea27ecedfd43dc1e
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Wed Jan 25 18:39:55 2023 +0100

    Revert "Revert "Revert "Revert "readme update""""

    This reverts commit a4cdffdf30e83dc51c84719f957a4e9191afe95e.

commit a4cdffdf30e83dc51c84719f957a4e9191afe95e
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Wed Jan 25 18:39:51 2023 +0100

    Revert "Revert "Revert "readme update"""

    This reverts commit 179c6822b0c84ca157a2ea954e5eaffe4993ea67.

commit 179c6822b0c84ca157a2ea954e5eaffe4993ea67
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Wed Jan 25 18:39:45 2023 +0100

    Revert "Revert "readme update""

    This reverts commit 0a0c65ce199069ab9083b4e8da56b31b50a57f1d.

commit 0a0c65ce199069ab9083b4e8da56b31b50a57f1d
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Wed Jan 25 18:38:54 2023 +0100

    Revert "readme update"

    This reverts commit 6e730da75c933f69e3661fa6b2d32dc48f0ea33b.

commit 6e730da75c933f69e3661fa6b2d32dc48f0ea33b
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Wed Jan 25 18:37:38 2023 +0100

    readme update

commit 1d97cd2bdfca634be894a52456e9a99fd30733b9
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Wed Jan 25 18:32:55 2023 +0100

    readme update

commit 65d9d78d6eec176d4202b2e2ef71efce86f35f2a
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Wed Jan 25 18:31:01 2023 +0100

    readme update

commit cd2fd9463e6efe441e68e0cb0d8a61c794057d27
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Wed Jan 25 18:28:53 2023 +0100

    readme update

commit cc92d2a20062909a4e1760cb8ec3c7bea6bf6ef4
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Wed Jan 25 18:27:00 2023 +0100

    readme update

commit 8e15d0b2e0e4d2c08674631b5f4c233b316a81ab
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Wed Jan 25 18:23:47 2023 +0100

    readme update

commit 7b2671eb46d292a9a86c424ea930903ba6272594
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Wed Jan 25 18:22:32 2023 +0100

    readme update

commit 394c57fafc22ee0e8975aaf3e6947b0ce4dda205
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Wed Jan 25 18:14:33 2023 +0100

    easier installation

commit 9c4c49019d0db6dac7efe9dd2e188f3d0c99b614
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Wed Jan 25 18:14:14 2023 +0100

    readme update

commit 724a411dc4c0fa7b969de988dd4834240a8068e1
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Wed Jan 25 17:34:47 2023 +0100

    better makefile and debugging window

commit b953b592f22b682391921aa978bd221f92fea3f5
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Wed Jan 25 17:08:35 2023 +0100

    debuggin mode now available with a keypress

commit a1caf5cccbb0f50981814bbd9841cb6d5bf46d80
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Wed Jan 25 17:00:06 2023 +0100

    ctrl+c now works always no matter the program state

commit 94dac09e74a3f93f3c106598e7fae09a571f45f2
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Wed Jan 25 16:19:45 2023 +0100

    Squashed commit of the following:

    commit 0373b72369b252eaab764dd9f9e24b517be10d62
    Author: P3rtang <tomverhulst21@gmail.com>
    Date:   Wed Jan 25 16:18:51 2023 +0100

        debug window early implementation

    commit 67ba9cf5509ada34f86b19e895470e379713e69c
    Author: P3rtang <tomverhulst21@gmail.com>
    Date:   Wed Jan 25 13:51:21 2023 +0100

        all features working again

    commit e658fbd094ec53f29d18e6d045922d5a5b4192d7
    Author: P3rtang <tomverhulst21@gmail.com>
    Date:   Wed Jan 25 00:25:11 2023 +0100

        early version of the new app state management

    commit fbc6311421a285b34ece6fc473569d4f39a092fd
    Author: P3rtang <tomverhulst21@gmail.com>
    Date:   Tue Jan 24 16:57:43 2023 +0100

        refactor appmodes to have an easy debugging option

commit df2be9f7294ff1626dee2f1f8dcb9a3d92ddbc69
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Tue Jan 24 00:02:01 2023 +0100

    small cleanup

commit 2a98d7e433ac22c35a5217ee3b34139824fc4198
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Mon Jan 23 23:32:02 2023 +0100

    implement input threading and easy switch between input methods

    > available input methods are reading from /dev/input if user is either
    part of the input group or the program is run as super user
    > the other input method is just the terminal input buffer
    > the input buffer runs concurently to the program and can be polled
    with a FiFo poll function
    > only keyevents are implemented for now

commit 1970351b5d042bfcda012bb07a49d626d4d99e94
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Sun Jan 22 20:41:16 2023 +0100

    changed how keylogging works

commit 6a68a51377ab93844107f2962ab3f55bfd455e98
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Sat Jan 21 18:14:26 2023 +0100

    clippy fixes

commit 1c12f271d2309e13c23bae664d4e127cd1f94116
Author: P3rtang <tomverhulst21@gmail.com>
Date:   Sat Jan 21 18:10:26 2023 +0100

    small fixes

+++ src/widgets/entry.rs
@@ -151,7 +151,7 @@ impl<'a> StatefulWidget for Entry<'a> {
         for i in 0..widget_area.height {
             buf.set_spans(
                 widget_area.x,
-                widget_area.y + i as u16,
+                widget_area.y + i,
                 &widget_empty.lines[0],
                 widget_area.width,
             );
